requirements.txt
streamlit
langchain
langchain_core
langchain_community
python-dotenv
groq
langchain-groq
ipykernel
faiss-cpu
pypdf